+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Treasure hunt".
 - Tests for "Self-driving car testing".
 - Tests for "No More Pythons, Please!".
+- Tests for "TicTacToe".
 
 ## [1.8.0] - 2022-04-30
 ### Added